removing directory
