+++ app/src/main/java/com/cookandroid/capstone_front_android/category_menu/category_display.java
@@ -1,11 +1,10 @@
-package community_menu;
+package com.cookandroid.capstone_front_android.category_menu;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -14,28 +13,28 @@ import androidx.fragment.app.Fragment;
 import com.cookandroid.capstone_front_android.MainActivity;
 import com.cookandroid.capstone_front_android.R;
 
-public class community_write extends Fragment {
+public class category_display extends Fragment {
     private View view;
 
     private MainActivity activity;
 
     private Button btn_exit; // 나가기 버튼.
 
-
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        view = inflater.inflate(R.layout.community_write, container, false);
+        view = inflater.inflate(R.layout.category_display, container, false);
 
         activity = (MainActivity) getActivity();
 
-        // 버튼설정.
         btn_exit = view.findViewById(R.id.exit);
 
         // 버튼이벤트.
         btn_exit.setOnClickListener(new View.OnClickListener(){
             @Override
-            public void onClick(View view) { activity.setCommunity(0); }
+            public void onClick(View view) {
+                activity.setCategory(0);
+            }
         });
 
         return view;
