+++ app/src/main/java/com/cookandroid/capstone_front_android/community_menu/community_review.java
@@ -1,11 +1,10 @@
-package community_menu;
+package com.cookandroid.capstone_front_android.community_menu;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -14,7 +13,7 @@ import androidx.fragment.app.Fragment;
 import com.cookandroid.capstone_front_android.MainActivity;
 import com.cookandroid.capstone_front_android.R;
 
-public class community_all extends Fragment {
+public class community_review extends Fragment {
     private View view;
 
     private MainActivity activity;
@@ -24,15 +23,11 @@ public class community_all extends Fragment {
     private Button btn_together; // 같이 갈 사람.
     private Button btn_review; // 리뷰.
     private Button btn_write; // 글쓰기.
-    
-    private ServiceApi service;
-    Retrofit retrofit;
-    ItemAdapter adapter;
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        view = inflater.inflate(R.layout.item_community_all,container,false);
+        view = inflater.inflate(R.layout.item_community_review,container,false);
 
         activity = (MainActivity) getActivity();
 
@@ -45,7 +40,9 @@ public class community_all extends Fragment {
         // 버튼이벤트.
         btn_all.setOnClickListener(new View.OnClickListener(){
             @Override
-            public void onClick(View view) { activity.setCommunity(0); }
+            public void onClick(View view) {
+                activity.setCommunity(0);
+            }
         });
 
         btn_together.setOnClickListener(new View.OnClickListener(){
@@ -64,19 +61,11 @@ public class community_all extends Fragment {
 
         btn_write.setOnClickListener(new View.OnClickListener(){
             @Override
-            public void onClick(View view) { activity.setCommunity(3); }
+            public void onClick(View view) {
+                activity.setCommunity(3);
+            }
         });
 
-        listView = (ListView) view.findViewById(R.id.list);
-
-        adapter = new ItemAdapter();
-        listView.setAdapter(adapter);
-
-        adapter.addItem("제목1", "내용1");
-        adapter.addItem("제목2", "내용2");
-        adapter.addItem("제목3", "내용3");
-        adapter.notifyDataSetChanged();
-        
         return view;
     }
 }
