+++ app/src/main/java/com/cookandroid/capstone_front_android/community_menu/community_write.java
@@ -1,4 +1,4 @@
-package category_menu;
+package com.cookandroid.capstone_front_android.community_menu;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -13,28 +13,28 @@ import androidx.fragment.app.Fragment;
 import com.cookandroid.capstone_front_android.MainActivity;
 import com.cookandroid.capstone_front_android.R;
 
-public class category_display extends Fragment {
+public class community_write extends Fragment {
     private View view;
 
     private MainActivity activity;
 
     private Button btn_exit; // 나가기 버튼.
 
+
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        view = inflater.inflate(R.layout.category_display, container, false);
+        view = inflater.inflate(R.layout.community_write, container, false);
 
         activity = (MainActivity) getActivity();
 
+        // 버튼설정.
         btn_exit = view.findViewById(R.id.exit);
 
         // 버튼이벤트.
         btn_exit.setOnClickListener(new View.OnClickListener(){
             @Override
-            public void onClick(View view) {
-                activity.setCategory(0);
-            }
+            public void onClick(View view) { activity.setCommunity(0); }
         });
 
         return view;
